基础功能重构

+++ src/main/java/com/example/backend/service/DiskService.java
@@ -1,7 +1,12 @@
 package com.example.backend.service;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.example.backend.entity.PodInfo;
+import com.example.backend.entity.VolumeInfo;
 import com.example.backend.k3s.disk.Disk;
+import com.example.backend.mapper.PodInfoMapper;
+import com.example.backend.mapper.VolumeInfoMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -14,11 +19,18 @@ import java.util.List;
 @Service
 public class DiskService {
 
-   @Value("${disk.path}")
+    @Value("${disk.path}")
     private String diskPath;
-//
+
+    @Autowired
+    private VolumeInfoMapper volumeInfoMapper;
+    //
 //   @Autowired
 //   private VolumeMapper volumeMapper;
+    @Autowired
+    private PodInfoMapper podInfoMapper;
+
+
 
     public boolean create(Disk disk) {
         // 在diskPath下创建一个名为disk.getName()的文件夹，如果存在则报错
@@ -44,12 +56,20 @@ public class DiskService {
         try {
             Process process = processBuilder.start();
             process.waitFor();
-//            Volume volume = new Volume();
-//            volume.setVolumeName(disk.getName());
-//            volume.setSizeMb((getDirectorySize(diskFile)/1024/1024));
-//            if(volumeMapper.insert(volume) <= 0){
-//                return false;
-//            }
+
+            VolumeInfo volumeInfo = new VolumeInfo();
+            volumeInfo.setVolumeName(disk.getName());
+            volumeInfo.setSizeMb((getDirectorySize(diskFile)/1024/1024));
+            volumeInfo.setPodName(disk.getPodName());
+            volumeInfo.setPodPath(disk.getPodPath());
+
+            LambdaQueryWrapper<PodInfo> wrapper= new LambdaQueryWrapper<>();
+            wrapper.eq(PodInfo::getPodName,disk.getPodName());
+            volumeInfo.setPodId(podInfoMapper.selectOne(wrapper).getPodId());
+
+            if(volumeInfoMapper.insert(volumeInfo) <= 0){
+                return false;
+            }
 //             output stdout and stderr
 //            BufferedReader reader = new BufferedReader(new java.io.InputStreamReader(process.getInputStream()));
 //            BufferedReader errorReader = new BufferedReader(new java.io.InputStreamReader(process.getErrorStream()));
@@ -62,11 +82,16 @@ public class DiskService {
 //                System.out.println(line);
 //            }
 //
+
         } catch (Exception e) {
             e.printStackTrace();
             return false;
         }
         return true;
+    }
+
+    public List<VolumeInfo> listAllVolumes() {
+        return volumeInfoMapper.selectList(null);
     }
 
     public boolean delete(Disk disk) {
@@ -76,11 +101,12 @@ public class DiskService {
 
         System.out.println(diskFile.exists()+" "+diskFile.isDirectory());
         boolean flag = diskFile.canRead();
-//        QueryWrapper<Volume> wrapper = new QueryWrapper<>();
-//        wrapper.eq("volume_name",disk.getName());
-//        if (!diskFile.getAbsoluteFile().exists() || volumeMapper.delete(wrapper) <= 0) {
-//            return false;
-//        }
+        QueryWrapper<VolumeInfo> wrapper = new QueryWrapper<>();
+        wrapper.eq("volume_name",disk.getName());
+        Integer res = volumeInfoMapper.delete(wrapper);
+        if (!diskFile.getAbsoluteFile().exists() || res <= 0) {
+            return false;
+        }
         return deleteAllFileInDirectory(diskFile.getAbsolutePath());
     }
 
@@ -157,4 +183,9 @@ public class DiskService {
         return size;
     }
 
+    public Object getDisk(String name) {
+        QueryWrapper<VolumeInfo> wrapper = new QueryWrapper<>();
+        wrapper.eq("volume_name",name);
+        return volumeInfoMapper.selectOne(wrapper);
+    }
 }
